refactor: 클래스 디렉터리 변경 HangmanInfo -> inner 클래스에서 domain으로 이동 MenuNumber, SingleAlphabet -> dto 에서 vo로 이동
